+++ Example/Pods/Target Support Files/Pods-SecureAccessBLE/Pods-SecureAccessBLE-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_SecureAccessBLE : NSObject
+@end
+@implementation PodsDummy_Pods_SecureAccessBLE
+@end