+++ Example/SecureAccessBLE/SecureAccessBLE.h
@@ -1,0 +1,19 @@
+//
+//  SecureAccessBLE.h
+//  SecureAccessBLE
+//
+//  Created by Oleg Langer on 20.10.17.
+//  Copyright © 2017 Huf Secure Mobile GmbH. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SecureAccessBLE.
+FOUNDATION_EXPORT double SecureAccessBLEVersionNumber;
+
+//! Project version string for SecureAccessBLE.
+FOUNDATION_EXPORT const unsigned char SecureAccessBLEVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SecureAccessBLE/PublicHeader.h>
+
+