+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- 2017-10-23 : (PLAM-1661)  : Added the new Cocoa Touch Framework target for binary distribution of the library   
+
 ### Changed
 ### Removed
 ### Fixed